+++ source/index.html
@@ -83,11 +83,11 @@
     <section class="screen screen--hidden screen--intro js-scrollspy" id="top">
       <div class="screen__wrapper">
         <div class="intro">
-          <h1 class="intro__title js-letter-animation letter-animation">Таинственный отпуск</h1>
+          <h1 class="intro__title"><div class="js-intro-title text">Таинственный</div> <div class="js-intro-title-two text">отпуск</div></h1>
           <div class="intro__content">
             <div class="intro__info">
-              <p class="intro__label js-letter-animation letter-animation">Даты проведения конкурса:</p>
-              <p class="intro__date">01 — 31.05 / 2020</p>
+              <p class="intro__label">Даты проведения конкурса:</p>
+              <p class="intro__date text js-intro-date">01 — 31.05 / 2020</p>
             </div>
             <div class="intro__message">
               <p>
@@ -136,7 +136,7 @@
         <div class="swiper-wrapper">
 
           <div class="slider__item swiper-slide">
-            <h2 class="slider__item-title">История</h2>
+            <h2 class="slider__item-title text js-history">История</h2>
             <p class="slider__item-text">Джеймс Таргет был основателем компании &laquo;Fjord Inc.&raquo;. С&nbsp;1965 года &laquo;Fjord Inc.&raquo; прославилась надёжными и&nbsp;удобными чемоданами и&nbsp;аксессуарами для туристов.</p>
           </div>
 
@@ -188,7 +188,7 @@
     <section class="screen screen--hidden screen--prizes js-scrollspy" id="prizes">
       <div class="screen__wrapper">
         <div class="prizes">
-          <h2 class="prizes__title">ПРизы</h2>
+          <h2 class="prizes__title text js-prizes">ПРизы</h2>
           <div class="prizes__lead">
             <p>Найди ответ и&nbsp;получи один из&nbsp;призов:</p>
           </div>
