+++ source/scss/blocks/animations.scss
@@ -1,0 +1,13 @@
+.text__word {
+    display: inline-block;
+    margin: 0 5px;
+    overflow: hidden;
+}
+.text__word > span {
+    transform: translate3d(0, 1em, 0);
+    display: inline-block;
+    padding-top: 2rem;
+}
+.text.active .text__word > span {
+    transform: translate3d(0, 0, 0);
+}
