+++ source/scss/blocks/intro.scss
@@ -135,6 +135,18 @@
     text-transform: uppercase;
     font-size: 2.4rem;
     line-height: 1.2;
+    animation: fadeinLabel 0.2s linear 1.5s;
+    animation-fill-mode: backwards;
+    @keyframes fadeinLabel {
+        0% {
+            opacity: 0;
+            transform: translateY(1rem);
+        }
+        100% {
+            opacity: 1;
+            transform: translateY(0);
+        }
+    }
     @media (min-width: $stop-scaling) {
         margin: 0 0 16px;
         width: 250px;
@@ -148,7 +160,7 @@
     font-weight: 400;
     font-size: 4.8rem;
     line-height: 0.906;
-    letter-spacing: -0.04em;
+    letter-spacing: 0.01em;
     @media (min-width: $stop-scaling) {
         margin: 0 0 0 -5px;
         font-size: 48px;
