+++ source/scss/style.scss
@@ -19,18 +19,4 @@
 @import "blocks/screen.scss";
 @import "blocks/slider.scss";
 @import "blocks/social-block.scss";
-
-.letter-animation-active {
-    display: inline-block;
-    overflow: hidden;
-}
-  
-.letter-animation-active > span {
-    transform: translate3d(0, 13rem, 0);
-    display: inline-block;
-}
-  
-.letter-animation.active .letter-animation-active > span {
-    transform: translate3d(0, 0, 0);
-    padding-top: 2rem;
-}
+@import "blocks/animations.scss";
